+++ guillotina/auth/groups.py
@@ -1,33 +1,34 @@
 from guillotina import configure
 from guillotina.auth.users import GuillotinaUser
 from guillotina.interfaces import IGroups
+from guillotina.interfaces import IPrincipal
+
+import typing
 
 
 class GuillotinaGroup(GuillotinaUser):
-    def __init__(self, ident):
-        super(GuillotinaGroup, self).__init__(ident)
-        self.id = ident
-
-        if ident == "Managers":
-            # Special Case its a Root Manager user
-            self._roles["guillotina.ContainerAdmin"] = 1
-            self._roles["guillotina.ContainerDeleter"] = 1
-            self._roles["guillotina.Owner"] = 1
-            self._roles["guillotina.Member"] = 1
-            self._roles["guillotina.Manager"] = 1
+    pass
 
 
 @configure.utility(provides=IGroups)
 class GroupsUtility:
     """ Class used to get groups. """
 
-    def get_principal(self, ident, principal=None):
-        if principal is not None:
-            try:
-                cache = principal._groups_cache
-            except AttributeError:
-                cache = principal._groups_cache = {}
-            if ident not in cache:
-                cache[ident] = GuillotinaGroup(ident)
-            return cache[ident]
-        return GuillotinaGroup(ident)
+    def __init__(self):
+        self._managers_group = GuillotinaGroup(
+            roles={
+                "guillotina.ContainerAdmin": 1,
+                "guillotina.ContainerDeleter": 1,
+                "guillotina.Owner": 1,
+                "guillotina.Member": 1,
+                "guillotina.Manager": 1,
+            }
+        )
+
+        self._groups: typing.Dict[str, GuillotinaGroup] = {"Managers": self._managers_group}
+
+    async def load_groups(self, groups):
+        pass
+
+    def get_principal(self, ident: str, principal: typing.Optional[IPrincipal]) -> IPrincipal:
+        return self._groups.get(ident) or GuillotinaGroup()