+++ guillotina/auth/users.py
@@ -12,6 +12,9 @@ class BaseUser:
     groups: list
     id: str
 
+    def __repr__(self):
+        return f"<{self.__class__.__name__}: {self.id}>"
+
 
 class SystemUser(BaseUser):
     id = "guillotina.SystemUser"