+++ guillotina/contrib/dbusers/__init__.py
@@ -5,7 +5,16 @@ from guillotina.i18n import MessageFactory
 _ = MessageFactory("guillotina.contrib.dbusers")
 
 
-app_settings = {"auth_user_identifiers": ["guillotina.contrib.dbusers.users.DBUserIdentifier"]}
+app_settings = {
+    "auth_user_identifiers": ["guillotina.contrib.dbusers.users.DBUserIdentifier"],
+    "load_utilities": {
+        "guillotina.dbusers": {
+            "provides": "guillotina.interfaces.IGroups",
+            "factory": "guillotina.contrib.dbusers.groups.DbUsersGroupsUtility",
+            "settings": {},
+        }
+    },
+}
 
 
 def includeme(root, settings):