+++ guillotina/tests/dbusers/test_security.py
@@ -8,32 +8,49 @@ import pytest
 
 pytestmark = pytest.mark.asyncio
 
+configure.permission("dbusers.SeeTopSecret", "SeeTopSecret", "Ability to see TopSecret docs")
+configure.role("dbusers.DoubleO", "00 Agent")
+configure.grant(permission="dbusers.SeeTopSecret", role="dbusers.DoubleO")
+
+
+@configure.service(
+    context=IFolder, method="GET", permission="dbusers.SeeTopSecret", name="@top-secret",
+)
+async def top_secret(context, request):
+    return {"documents": ["abcd"]}
+
 
 @pytest.mark.app_settings(settings.DEFAULT_SETTINGS)
 async def test_roles_in_groups(dbusers_requester):
-    configure.permission("dbusers.SeeTopSecret", "SeeTopSecret", "Ability to see TopSecret docs")
-    configure.role("dbusers.DoubleO", "00 Agent")
-    configure.grant(permission="dbusers.SeeTopSecret", role="dbusers.DoubleO")
-
-    @configure.service(
-        context=IFolder, method="GET", permission="dbusers.SeeTopSecret", name="@top-secret",
-    )
-    async def top_secret(context, request):
-        return {"documents": ["abcd"]}
 
     async with dbusers_requester as requester:
+        # Create the group 'top-agents' and assign the role "dbusers.DoubleO"
         resp, status = await requester(
             "POST",
             "/db/guillotina/groups",
-            data=json.dumps({"id": "top-agents", "@type": "Group", "user_roles": ["dbusers.DoubleO"],}),
+            data=json.dumps({"id": "top-agents", "@type": "Group", "user_roles": ["dbusers.DoubleO"]}),
         )
         assert status == 201
-        resp, status = await requester("GET", "/db/guillotina/@groups")
-        assert status == 200
-        assert len(resp) == 1
-        assert resp[0]["roles"] == ["dbusers.DoubleO"]
 
-        # create the user
+        # Grant guillotina.AccessAcontent to role 'dbusers.DobuleO'
+        _, status = await requester(
+            "POST",
+            "/db/guillotina/@sharing",
+            data=json.dumps(
+                {
+                    "roleperm": [
+                        {
+                            "role": "dbusers.DoubleO",
+                            "permission": "guillotina.AccessContent",
+                            "setting": "Allow",
+                        }
+                    ]
+                }
+            ),
+        )
+        assert status == 200
+
+        # Create the user and assign it to group 'top-agents'
         resp, status = await requester(
             "POST",
             "/db/guillotina/users",
@@ -50,30 +67,30 @@ async def test_roles_in_groups(dbusers_requester):
         )
         assert status == 201
 
+        # Check the group 'top-agents' to ensure it contains the user
         resp, status = await requester("GET", "/db/guillotina/@groups/top-agents")
         assert status == 200
         assert resp["users"]["items"] == ["007"]
 
-        # Create folder 'secret's
+        # Create folder 'secrets'
         resp, status = await requester(
             "POST", "/db/guillotina/", data=json.dumps({"@type": "Folder", "id": "secrets"}),
         )
         assert status == 201
 
-        # resp, status = await requester(
-        #     "GET",
-        #     "/db/guillotina/secrets/@top-secret",
-        # )
-        # assert status == 404
+        # Check the default user can't access the endpoint
+        # (doesn't have permission 'dbusers.SeeTopSecret')
+        resp, status = await requester("GET", "/db/guillotina/secrets/@top-secret",)
+        assert status == 401
 
-        # Login as new user
-        resp, status_code = await requester(
+        # Login as James Bond
+        resp, status = await requester(
             "POST", "/db/guillotina/@login", data=json.dumps({"username": "007", "password": "secret"})
         )
-        assert status_code == 200
+        assert status == 200
 
-        # User should have access to its own folder
-        _, status_code = await requester(
+        # Check that James Bond has access to the Top Secret documents
+        _, status = await requester(
             "GET", "/db/guillotina/secrets/@top-secret", token=resp["token"], auth_type="Bearer"
         )
-        assert status_code == 200
+        assert status == 200