+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 6.0.0a8 (unreleased)
 --------------------
 
+- Improvements in contrib.dbusers
+  [masipcat]
 
 - Fix command 'create'
   [masipcat]